+++ CHANGELOG.md
@@ -1,9 +1,23 @@
-### 0.0.1
+## [0.0.2] - 2025-10-15
 
-** Initial **
+### Changed
 
-  - Initial Release.
+- rngCNPJfiliais: possibility of informing the desired branch (1 to 9999) when generating a CNPJ
 
+### Added
+
+- valBissexto: Check if it is a leap year
+- dtPascoa: Calculates the date of Easter Sunday for a given year
+- valPascoa: Check if a specific date is Easter Sunday
+- dtCarnaval: Calculates the date of Carnival Tuesday for a given year
+- valCarnaval: Check if a specific date is Carnival Tuesday
+- ufCep: Check a zip code and return a list containing city and state
+
+## [0.0.1] - 2025-10-12
+
+### Added
+
+- Initial Release.